Backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
huggingface-hub==0.27.1
python-dotenv==1.0.1
